Refactored code for easy pytest integration
