+++ pyproject.toml
@@ -22,6 +22,7 @@ tensorflow-io-gcs-filesystem = [
 ]
 ml-lib-remla = "^1.0.0"
 pylint = "^3.2.2"
+pytest = "^8.2.1"
 
 [build-system]
 requires = ["poetry-core"]
