+++ reports/metrics.json
@@ -1,1 +1,1 @@
-{"train_accuracy": 0.5084999799728394, "train_loss": 0.6965246796607971, "val_accuracy": 0.45010000467300415, "val_loss": 0.6957074999809265, "test_accuracy": 0.5584, "avg_precision": 0.6040955631399317, "avg_recall": 0.039535403171766806, "avg_f1": 0.07421383647798742, "roc_auc": 0.509266162567234}
+{"train_accuracy": 0.5084999799728394, "train_loss": 0.6965246796607971, "val_accuracy": 0.45010000467300415, "val_loss": 0.6957074999809265, "test_accuracy": 0.4597, "avg_precision": 0.41425925925925927, "avg_recall": 0.49966495421040874, "avg_f1": 0.45297155006580947, "roc_auc": 0.46348447782944846}
